changes to landing page
